backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.0
django-cors-headers==4.3.1
cryptography==42.0.5
python-decouple==3.8
supabase==2.3.4 
PyJWT==2.10.1
requests==2.32.4